### 2018-01-10 #185 @kvantas
Submission: hydroscoper
*hydroscoper* is an interface to the Greek National Data Bank for Hydrological and Meteorological Information, Hydroscope. It provides functions to: a) Transform the available tables and data sets into tidy data frames. b) Transliterate the Greek Unicode text to Latin. c) Translate various Greek terms to English.

```
Package: hydroscoper
Title: Interface to Hydroscope
Version: 0.2.1
Authors@R: person("Konstantinos", "Vantas", role = c("aut", "cre"),
  email = "kon.vantas@gmail.com",
Maintainer: Konstantinos Vantas <kon.vantas@gmail.com>
Description: R interface to the  Greek National Data Bank for Hydrological and
    Meteorological Information <http://www.hydroscope.gr/>. It
    covers Hydroscope's data sources and provides functions to transliterate,
    translate and download them into tidy dataframes (tibbles).
Imports: stringi (>= 1.1.6),
    stringr (>= 1.2),
    tibble(>= 1.4.1),
    readr (>= 1.1.1),
    jsonlite (>= 1.5),
Suggests: plyr (>= 1.8.4),
    ggplot2 (>= 2.2.1),
    knitr (>= 1.17),
    rmarkdown (>= 1.8),
    ggmap (>= 2.6.1)
```

https://github.com/kvantas/hydroscoper

data retrieval, because the package parses and download data from a national data-bank.

a) Engineers for the development of water resources and environmental studies in Greece, b) scientists  that work with hydrological and meteorological data from Hydroscope and c) Greek organizations to submit data and reports to the European Union for the Implementation of the Water Framework Directive.

No other package exists.

- [x] Do you intend for this package to go on CRAN?  (allready on CRAN)
    - [x] The package is deposited in a long-term repository with the DOI:  10.5281/zenodo.1120431
	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

```
R CMD check results
0 errors | 0 warnings | 0 notes
```
