### 2018-01-10 #185 @kvantas
using cover::package_coverage() using local tests (i.e. comment line 20 in file /test/test_get.R) I get:
```
hydroscoper Coverage: 98.94%
R/get_data.R: 90.91%
R/enhydris.R: 98.04%
R/defunct.R: 100.00%
R/get_tables.R: 100.00%
R/hydro_coords.R: 100.00%
R/hydro_translate.R: 100.00%
R/translit.R: 100.00%
```

and with goodpractice::gp():
```
It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 98% of code
    lines are covered by test cases.

    R/enhydris.R:77:NA
    R/get_data.R:64:NA
```
