### 2018-01-11 #185 @maelle
@kvantas I get an error when checking the package

```r
ERROR
Quitting from lines 28-38 (hydroscope_data.Rmd)
Error: processing vignette 'hydroscope_data.Rmd' failed with diagnostics:
Failed to parse url: http://deh.hydroscope.gr/api/Station/?format=json
```

I cannot open this URL in the browser either. If the website often fails, maybe you should add some error handling?

I've got two reviewers lined up so as soon as you fix this, the review process can begin! 😉
