### 2018-01-11 #185 @maelle
The problem is with `get_database("deh")`. The call to `get_database` works for the other subdomains.
