### 2018-01-11 #185 @kvantas
@maelle, thanks for the feedback.

I suspect that the server is down for maintenance,  to upgrade the Enhydris database system and the site's front-end. `deh.hydroscope.gr` used older versions of both.  I am using Hydroscope the last two years and its sub-domains rarely fail.

I sent an email to the Special Secretariat for Water for info.  Do you think that I should add error handling anyway?
