### 2018-01-11 #185 @maelle
Thanks for your quick answer. Yes it might give a better error message to the user, so that they know what's going on. Please update this thread once you know the server is up again!
