+++ DESCRIPTION
@@ -27,5 +27,6 @@ Suggests: plyr (>= 1.8.4),
     knitr (>= 1.17),
     rmarkdown (>= 1.8),
     testthat (>= 1.0.2),
+    pingr (>= 1.1.2),
     ggmap (>= 2.6.1)
 VignetteBuilder: knitr
