### 2018-01-12 743a7eb
add nolint for subdomain addresses
