### 2018-01-12 205591d
use conditional testing on sub-domain availability
check if expected tables exist in subdomains

remove test about variable names in data
