+++ tests/testthat/test_get.R
@@ -16,76 +16,92 @@ test_that("Missing arguments and unknown subdomains return errors", {
   expect_error(get_data("kyy"))
 })
 
-# comment the following line for local testing
-skip("heavy web usage")
+# To test local using devtools::test() comment the following line
+skip("Heavy web usage")
 
 context("Test API")
-test_that("API returns a list of tables", {
-  expect_is( get_api_json("kyy"), "list")
-  expect_is( get_api_json("ypaat"), "list")
-  expect_is( get_api_json("emy"), "list")
-  expect_is( get_api_json("deh"), "list")
-})
+test_that("APIs return lists of expected tables", {
 
-context("Check if data from hydroscope are as expected")
-test_that("get functions return dataframes with a subset of variables", {
-
-  # get data from subdomains ---------------------------------------------------
-  expect_is(kyy <- get_database("kyy"), "list")
-  expect_is(emy <- get_database("emy"), "list")
-  expect_is(ypaat <- get_database("ypaat"), "list")
-  expect_is(deh <- get_database("deh"), "list")
-
-  # dataframes have expected variables -----------------------------------------
-  expect_have_vars <- function(df, vars) {
-    eval(bquote(expect_true(all(.(vars) %in% names(.(df))))))
+  # create booleans if servers are online
+  online <- function(h_url){
+    !is.na(pingr::ping(h_url, count = 1))
   }
 
-  expect_database <- function(subdomain, st_vars, ts_vars, wb_vars,
-                              wd_vars, var_vars, own_vars, unit_vars,
-                              tstep_vars) {
-    # stations
-    eval(bquote(expect_have_vars(.(subdomain)$stations, st_vars)))
-    # timeseries
-    eval(bquote(expect_have_vars(.(subdomain)$timeseries, ts_vars)))
-    # water_basins
-    eval(bquote(expect_have_vars(.(subdomain)$water_basins, wb_vars)))
-    # water_divisions
-    eval(bquote(expect_have_vars(.(subdomain)$water_divisions, wd_vars)))
-    # variables
-    eval(bquote(expect_have_vars(.(subdomain)$variables, var_vars)))
-    # owners
-    eval(bquote(expect_have_vars(.(subdomain)$owners, own_vars)))
-    # units
-    eval(bquote(expect_have_vars(.(subdomain)$units, unit_vars)))
-    # time_steps
-    eval(bquote(expect_have_vars(.(subdomain)$time_steps, tstep_vars)))
+  kyy   <- online("kyy.hydroscope.gr")
+  emy   <- online("emy.hydroscope.gr")
+  ypaat <- online("ypaat.hydroscope.gr")
+  deh   <- online("deh.hydroscope.gr")
+
+
+  exp_names <- c("Station", "Timeseries", "Instrument", "WaterBasin",
+                "WaterDivision", "PoliticalDivision", "Variable",
+                "UnitOfMeasurement", "TimeStep", "Organization",
+                "InstrumentType", "StationType")
+
+
+  expect_tables_in_api <- function(subomain) {
+    expect_true(all(exp_names %in% names(get_api_json(subomain))))
   }
 
-  st_vars <- c("id", "name", "owner", "point", "altitude", "water_basin",
-                     "water_division")
-  wd_vars <- wb_vars <- own_vars <- c("id", "name")
-  ts_vars <- c("id", "gentity", "start_date_utc", "end_date_utc", "variable",
-                "time_step", "unit_of_measurement")
-  var_vars <- tstep_vars <- c("id", "descr")
-  unit_vars <- c("id", "symbol")
[…]
