### 2018-01-12 464b5c8
rewrite get_ tests
