+++ README.Rmd
@@ -56,8 +56,8 @@ The data are structured as tables and space separated text files, but are in Gre
 Note that:
 
 1. Only the two Ministries allow to download time series values freely.
-2. `ypaat`, `emy` and `kyy` subdomains are maintened by the National Technical University Of Athens and these servers work seamlessly.
-3. `deh` subdomain is maintened by the Greek Public Power Corporation, uses an older version of the Enhydris API and occasionally the server is down.
+2. `ypaat`, `emy` and `kyy` subdomains are maintained  by the National Technical University Of Athens and these servers work seamlessly.
+3. `deh` subdomain is maintained  by the Greek Public Power Corporation, uses an older version of the Enhydris API and occasionally the server is down.
 
  
 ## Installation
