### 2018-01-12 #185 @kvantas
Hydroscope reported to me that `ypaat`, `emy` and `kyy` sub-domains are maintained by the National Technical University Of Athens and that `deh` sub-domain is maintained by the Greek Public Power Corporation and  that this server was down for some days three months ago.

I excluded the use of `deh` in the vignette and I added conditionally evaluation for the chunks in the Rmd file based on the availability of the sub-domains and the tibbles' variables that the APIs return. I  think (and hope) that with this way this vignette should never fail.

devtools::check() gives me again
```
R CMD check results
0 errors | 0 warnings | 0 notes
```
thanks ;)
