### 2018-01-12 #185 @maelle
Ok, thanks! I still think that you should add more error handling, so that if it fails, an user trying to acces `deh` might get something like "The server for that data source is probably down, get more info at blablabla or try again later".

Will update this thread with editor checks etc a bit later!
