### 2018-01-12 #185 @maelle
### Editor checks:

---

#### Editor comments

Thanks for your submission, @kvantas! :smile_cat: My only note for you and the reviewers is the error handling of that `deh` subdomain. I'll erase a few comments to keep the thread clean.

---

Reviewers: @timtrice @theoniphotopoulou
Due date: 2018-02-04
