### 2018-01-12 #185 @maelle
@timtrice @theoniphotopoulou thanks for accepting to review this package! 😄

Your reviews are due on 2018-02-04.

Please find here [our reviewing guide](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md) and [the review template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md)
