### 2018-01-12 ad49898
enrich error handling
