+++ R/get_data.R
@@ -49,13 +49,16 @@ get_data <- function(subdomain =  c("kyy", "ypaat"), time_id) {
   # check that stationID is given
   if (is.null(time_id)) stop("argument \"time_id\" is missing")
 
+  # check if a sudomain is down
+  down_server(subdomain)
+
   # match subdomain values
   subdomain <- match.arg(subdomain)
   h_url <- hydroscope_url(subdomain)
   t_url <- paste0(h_url, "/api/tsdata/", time_id, "/") # nolint
 
-  # get hydroscope file to dataframe
 
+  # get hydroscope file to dataframe
   tryCatch({
     enhy_get_txt(t_url)
   },
