### 2018-01-12 #185 @kvantas
Following @maelle 's advice I enriched error handling about all the sub-domains availability. Thank you very much!
