+++ tests/testthat/test_get.R
@@ -17,7 +17,7 @@ test_that("Missing arguments and unknown subdomains return errors", {
 })
 
 # To test local using devtools::test() comment the following line
-# skip("Heavy web usage")
+skip("Heavy web usage")
 
 skip_if_not_online <- function(subdomain) {
   h_url <- paste0(subdomain, ".hydroscope.gr")
