### 2018-01-31 #185 @maelle
@kvantas for info one of the two reviewers had to drop out this time, so the second reviewer will be @sharlagelfand.

@sharlagelfand, thanks for accepting to review this package! Your review is due on 2018-02-21. 😸

Please find here [our reviewing guide](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md) and [the review template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md).
