### 2018-02-07 #185 @timtrice
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

1. Failed to parse URL (from vignette):

```r
library(hydroscoper)
library(plyr)
subdomain <- "kyy"
stations <- get_stations(subdomain)
coords  <- ldply(stations$StationID, function(id) {get_coords(subdomain, id)})
```

```
1: Failed to parse url: http://kyy.hydroscope.gr/stations/d/501032/

2: Failed to parse url: http://kyy.hydroscope.gr/stations/d/200425/

3: Failed to parse url: http://kyy.hydroscope.gr/stations/d/10014/
```

```r
timeseries <- ldply(stations$StationID, function(id) {get_timeseries(subdomain, id)})
```

```
1: Failed to parse url: http://kyy.hydroscope.gr/stations/d/10003/

2: Failed to parse url: http://kyy.hydroscope.gr/stations/d/10004/

3: Failed to parse url: http://kyy.hydroscope.gr/stations/d/10005/

4: Failed to parse url: http://kyy.hydroscope.gr/stations/d/10006/

5: Failed to parse url: http://kyy.hydroscope.gr/stations/d/10014/
```

3. Should contribution guidelines be named *CONTRIBUTING.md* rather than *CONDUCT.md*?

4. I am unsure about the JOSS questions with the exception that the package does **not** have a `paper.md` file as listed

5. I would recommend using the @title, @description, @details and @section tags in function documentation (just as @param was used). IMO, this improves legibility.

6. Install from GitHub did not install all package documentation (missing several of the get* variables). Perhaps the documentation hasn't been fully updated in the master branch?

7. The tests seem to fail if the `skip("Heavy web usage")` line is commented,

> Exited with status 15.

I thought this had worked earlier but I may have been mistaken. If I leave it uncommented as it was then the tests pass.

8. `goodpractice::gp()` output:

── GP hydroscoper ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

    R/enhydris.R:19:NA
    R/enhydris.R:20:NA
    R/enhydris.R:26:NA
    R/enhydris.R:34:NA
    R/enhydris.R:42:NA

  ✖ fix this R CMD check NOTE: Note: found 3031 marked UTF-8 strings
