### 2018-02-08 #185 @kvantas
Many thanks, @timtrice,  for your time and effort! I will start working on these issues as soon as possible.
