### 2018-02-08 #185 @maelle
Thank you @timtrice for your review! 😸 And thanks @kvantas for being so reactive! Feel free to wait for the second review if that's easier for you. 😺

CONTRIBUTING.md and CONDUCT.md are two separate files that should both exist.

@kvantas do you plan a JOSS submission? If so, could you add a paper.md conform to JOSS guidelines? Thanks.
