### 2018-02-08 #185 @maelle
Oops actually sorry we do _not_ recommend any CONTRIBUTING.md yet!
