### 2018-02-08 #185 @maelle
If you solve bugs @kvantas please update this thread so that @sharlagelfand might know which package version to use for her review.
