### 2018-02-08 #185 @kvantas
That's OK, I can add CONTRIBUTING.md.

@maelle,  I already have a paper.md in inst/paper. I think that @timtrice used an older version of my package and not `0.2.1`, as I write in **`Summary`.**

Anyway, I will follow  @timtrice comments and update this thread about the package version.

Thanks again, and sorry if I am missing something obvious.
