### 2018-02-11 0026e79
fix review comments
1. Write a new vignette that uses internal data-sets
2. make error messages more user friendly
3. add CONTRIBUTING.md
4. edit and move paper.md to root
5. use @title, @description, @details and @Section  tags
6.  almost all get* functions use @rdname get_tables
7. rewrite tests using skip_on_cran() and a custom skip_if_not_online() function
8.  run goodpractice::gp() with clear output
