+++ paper.bib
@@ -34,6 +34,7 @@
 
 @article{peng2011reproducible,
   title={Reproducible research in computational science},
+  Doi = {10.1126/science.1213847},
   author={Peng, Roger D},
   journal={Science},
   volume={334},
@@ -71,6 +72,7 @@
 @article{kandel2011research,
   title={Research directions in data wrangling: Visualizations and transformations for usable and credible data},
   author={Kandel, Sean and Heer, Jeffrey and Plaisant, Catherine and Kennedy, Jessie and van Ham, Frank and Riche, Nathalie Henry and Weaver, Chris and Lee, Bongshin and Brodbeck, Dominique and Buono, Paolo},
+  Doi = {10.1177/1473871611415994},
   journal={Information Visualization},
   volume={10},
   number={4},
