+++ tests/testthat/test_defunct.R
@@ -1,0 +1,5 @@
+context("defunct related tests")
+
+test_that("get_coords returns an error", {
+  expect_error(get_coords())
+})

[…]
