### 2018-02-11 #185 @kvantas
Hello all, this is a list of the updates/changes I have made in response to  @timtrice 's review.

-------------
1., 2.  I replaced that vignette with a new one: `stations_with_data.Rmd` that uses the package's data-sets and does not make heavy web usage as the previous one.
3. I added `CONTRIBUTING.md`
4. I moved `paper.md` and `paper.bib` from inst/paper to the root directory .
5. I added  @title, @description, @details and @Section tags in function documentation.
6. I used @rdname to document  multiple functions (almost all the `get_...` functions).
7. I  tide up the internal, API-related functions, wrote tests again utilizing`testthat::skip_on_cran()` and a custom function, `skip_if_not_online()`; there is no need to comment anything in tests now.
8. Using `Sys.setenv(NOT_CRAN="true")` and `goodpractice::gp()` :
```
♥ Yee-haw! Classy package! Keep up the awe-inspiring work!
```
-------------
`Sys.setenv(NOT_CRAN="true")` and  `covr::package_coverage()` returns:
```
hydroscoper Coverage: 100.00%
```
`Sys.setenv(NOT_CRAN="false")` and  `devtools::check()` returns:
```
 0 errors ✔ | 0 warnings ✔ | 0 notes ✔
```
-------------

@timtrice, thanks for your review,  it really improved my package's structure and testing. Please let me know if I missed something.

@maelle  and @sharlagelfand  please use the package version `0.2.3`.
