### 2018-02-13 cb7685e
Typo fixes
