+++ README.Rmd
@@ -28,7 +28,7 @@ knitr::opts_chunk$set(
 
 `hydroscoper` is an R interface to the  Greek National Data Bank for Hydrological and Meteorological Information, [Hydroscope](http://www.hydroscope.gr/). Hydroscope is the result of long-standing efforts by numerous Greek scientists in collaboration with various companies and associations. It was implemented in three phases, funded by the Ministry of Development, the Ministry of Environment and Energy and the European Union.
 
-Hydroscope, provides several national data sources from various organisations via a web interface. Each participating organisation keeps its data on its own server using the Enhydris database system for the storage and management of hydrological and meteorological data. These organisations are:
+Hydroscope provides several national data sources from various organisations via a web interface. Each participating organisation keeps its data on its own server using the Enhydris database system for the storage and management of hydrological and meteorological data. These organisations are:
 
 * Ministry of Environment and Energy.
 * Ministry of Rural Development and Food.
@@ -128,7 +128,7 @@ station_ts
 We can download the station's  precipitation time series **56** (http://kyy.hydroscope.gr/timeseries/d/56/):
 
 ``` {r get_ts}
-ts_raw <- get_data(subdomain = "kyy",time_id = 56)
+ts_raw <- get_data(subdomain = "kyy", time_id = 56)
 ts_raw
 ```
 
