### 2018-02-18 #185 @sharlagelfand
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

I really enjoyed reviewing this package and especially from an end-user standpoint -- I hope you don't mind the specificness of some of these comments with that in mind! :)

```
library(covr)
Sys.setenv(NOT_CRAN="true")
covr::package_coverage()
```
```
hydroscoper Coverage: 100.00%
R/defunct.R: 100.00%
R/enhydris_api.R: 100.00%
R/get_data.R: 100.00%
R/get_tables.R: 100.00%
R/hydro_coords.R: 100.00%
R/hydro_translate.R: 100.00%
R/translit.R: 100.00%
R/utils.R: 100.00%
```

```
library(goodpractice)
goodpractice::gp()
```
```

✖ fix this R CMD check NOTE: Note: found 3031 marked UTF-8 strings
✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates
Rd problems. LaTeX errors found: ! LaTeX Error: File `inconsolata.sty' not found. Type X to quit or
<RETURN> to proceed, or enter new name. (Default extension: sty) ! Emergency stop. <read *> l.276 ^^M !
==> Fatal error occurred, no output PDF file produced!
```

This second one just might be my system -- I don't think I have LaTeX on this computer.

```
devtools::check()
R CMD check results
0 errors | 0 warnings | 0 notes

devtools::test()
✔ | OK F W S | Context
✔ |  1       | defunct related tests
✔ |  1       | Test enhydris_api helper functions
✔ |  2       | enhydris_url returns expected addresses
✔ |  3     1 | Test that expected tables exist in sub-domains' databases [11.6 s]
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
test_enhydris.R:62: skip: deh returns expected tables
deh.hydroscope.gr is not online
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |  4       | Test enhydris_get and enhydris_list functions⠋ |  1       | Test enhydris_get and enhydris_list functions [21.9 s]
✔ |  1       | Test enhydris_list function [2.5 s]
✔ |  2       | get_data function tests
✔ | 13       | get_tables family function tests
✔ |  2       | hydro_coords tests [0.1 s]
✔ |  7       | Translitetation and translation functions
⠋ |  1       | utils testsping: cannot resolve xxx.hydroscope.gr: Unknown host
✔ |  2       | utils tests

══ Results ════════════════════════════════

OK:       38

You are a coding rockstar!
```

Looks like the skipped test is expected, or it is expected to return an error, at least -- do you mind clarifying?

#### Examples run unsuccessfully
The example for `get_instruments_type()` fails. This is because it is not exported properly -- in `R/get_tables.R` I can see it has `#' @export get_owners` above it instead of `#' @export get_instruments_type` -- I assume once the correct function is exported, it will run fine (since the example for `get_database` runs just fine!).

The README is great. It provides good background on the Hydroscope project itself, what is included in it, and limitations (i.e., that the files are in Greek). I liked the list of what `hydroscoper` does with data from Hydroscope, and especially how it addresses the limitations to make the data more usable.

Something that did confuse me is the `timeseries` data set and reference to it. Perhaps it's that I'm missing the context, but I think further explanation of what this data set is would be helpful in both the README, the data set's documentation, and the vignettes. My impression is that it is a lookup table of all of the available measurements for a given station in a given domain, with units of measurement and the time that those measurements are for (hence the name `timeseries`) -- I think highlighting more concretely what exactly it is a time series of would be useful.

I think the vignette(s?) could use a gentler introduction to Hydroscope and the `hydroscoper` package, similar to the README. I'm imagining the case where someone finds this package on CRAN, not on GitHub, and needs more background into what the package has (functions, data sets) than is included in the current vignettes.

I'm not entirely clear if the `hydroscope_data` vignette is in the wrong place (`data-raw/` instead of `vignettes/`), or if it is meant to be used to create the `timeseries` data set (hence its location) and is not actually a vignette at all! I think the latter, since you created the `stations_with_data.Rmd` after @timtrice's review, but just wanted to double check :)

#### Function Documentation
* `get_data`
	* The documentation references an argument `timeID` when it is actually `time_id`
	* There is inconsistent usage of tibble/data frame/dataframe throughout
	* It says that it will return a data frame if the subdomain is "kyy" or "ypaat" and the time argument is not NULL, and an error otherwise; but it will also throw an error if the `time_id` is not NULL **and** it doesn't correspond to a timeseries in the `timeseries` lookup -- I wonder if it would be useful to be clearer that the subdomain and time_id combination must exist in `timeseries` (and therefore must exist in Hydroscope). It might sound silly, but the `time_id` not being NULL is not enough on its own.
	* The actual code for this function has a comment saying it is checking that `stationID` is not null, though it is actually checking that `time_id` is not null.

#### Miscellanious thoughts

* The `timeseries` data set has a column `timeser_id`, which is then used as an argument in `get_data()` -- however, the argument in this function is named `time_id`. I think it would be clearer to either rename the variable name in `timeseries` to `time_id`, or to change the argument in `get_data()` to `timeser_id`, so that they are consistent.
* The list that `get_database()` has an element called `units`, though the function that retrieves it is called `get_units_of_measurement()` -- all of the other list elements match the name of their corresponding functions, so I think it would be useful if this one did as well.
* The `greece_borders` data set is a `data.frame`, not a `tbl_df` like the other two data sets (the documentation indicates that it is a tibble as well).
* IMO the `get_data()` function should return columns `date`, `value`, and `comment`, rather than `dates`, `values`, and `comments` --  a single row contains the measurement (value) on a given date, with a comment. I don't think (?) it's customary for variable names to be pluralized.

I have also submitted a PR with a few fixes on documentation/typos -- nothing major, just when I saw things that could quickly be fixed and didn't need to be brought up here!
