### 2018-02-21 af0297e
#2 fix example for get_instruments_type()
export properly get_instruments_type in
get_tables and NAMESPACE
