+++ NAMESPACE
@@ -4,6 +4,7 @@ export(get_coords)
 export(get_data)
 export(get_database)
 export(get_instruments)
+export(get_instruments_type)
 export(get_owners)
 export(get_political_divisions)
 export(get_station_type)
