### 2018-02-21 d7472f3
add osx tests
