+++ .travis.yml
@@ -1,5 +1,9 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
+os:
+  - linux
+  - osx
+
 language: R
 sudo: false
 cache: packages
