### 2018-02-21 445f0cc
#3 get_data documentation
1. replace timeID with time_id
2. rewrite @return text
3. replace stationID with time_id at a comment
