### 2018-02-21 a12b06a
#6 get_data() should return varables date, value, and comment
1. edit get_data() documentation
2. edit plot aes() parameters
3. edit enhydris_txt()
