### 2018-02-21 59f4acb
#7 get_database() has an element called `units`, though the function that retrieves it is called `get_units_of_measurement`()
1. Edit the named list elenent `units`
