+++ R/get_tables.R
@@ -219,7 +219,7 @@ get_database <- function(subdomain =  c("kyy", "ypaat", "emy", "deh"),
        "water_divisions" = get_water_divisions(subdomain, translit),
        "political_divisions" = get_political_divisions(subdomain, translit),
        "variables" = get_variables(subdomain, translit),
-       "units" = get_units_of_measurement(subdomain, translit),
+       "units_of_measurement" = get_units_of_measurement(subdomain, translit),
        "time_steps" = get_time_steps(subdomain, translit),
        "owners" = get_owners(subdomain, translit),
        "instr_type" = get_instruments_type(subdomain, translit),
