### 2018-02-21 fdec5c5
#8 timeseries data set has a column `timeser_id`
1. edit data.R
2. rename timeseries variable to `time_id`
3. edit create_pck_data.R
