+++ R/data.R
@@ -3,7 +3,7 @@
 #' @description  Stations' data from the Greek National Data Bank for
 #' Hydrological and Meteorological Information.
 #'
-#' @format A data frame:
+#' @format A tibble with 2,322 rows and 9 variables:
 #' \describe{
 #'     \item{station_id}{The station's ID from the domain's database}
 #'     \item{name}{The station's name}
@@ -22,9 +22,9 @@
 #' @description Time series' data from the Greek National Data Bank for
 #' Hydrological and Meteorological Information.
 #'
-#' @format A data frame:
+#' @format A tibble with 10,804 rows and 9 variables:
 #' \describe{
-#'     \item{timeser_id}{The time series ID from the database}
+#'     \item{time_id}{The time series ID from the database}
 #'     \item{station_id}{The corresponding station's ID}
 #'     \item{variable}{The time series variable type}
 #'     \item{timestep}{The timestep of time series}
@@ -41,7 +41,7 @@
 #' variables are created using the function tidy from the broom package. This
 #' data frame was created for use with the geom_polygon from ggplot2 package.
 #'
-#' @format A data frame with 18474 rows and 8 variables:
+#' @format A tibble with 18474 rows and 8 variables:
 #' \describe{
 #'   \item{long}{Longitude in decimal degrees, ETRS89}
 #'   \item{lat}{Latitude in decimal degrees, ETRS89}
