### 2018-02-21 692bf20
follow functions' names in get_database()
