+++ paper.md
@@ -25,9 +25,9 @@ affiliations:
 
 The Greek National Data bank for Hydrological and Meteorological Information, Hydroscope [@hydroscope2018], is the result of long-standing efforts by numerous Greek scientists in collaboration with various companies and associations. Its main purpose is the formation of the basic infrastructure for the implementation of the European Community Directives 2000/60/EC, about the establishing of a framework for Community action in the field of water policy and 2007/60/EC, about the assessment and management of flood risks.
 
-Hydroscope provides several national data sources in HTML and plain text files, via a web interface, using the Enhydris database system [@christofides2011enhydris]. These data are well structured but are in Greek, thus limiting their usefulness. Fully reproducible research [@peng2011reproducible], can be tedious and error-prone using Hydroscope's web interface. On the contrary, using the Enhydris API for reproducibility requires external programs and scripting to import the data.
+Hydroscope provides several national data sources in HTML and plain text files, via a web interface, using the Enhydris database system [@christofides2011enhydris]. These data are well structured but are in Greek, thus limiting their usefulness. Fully reproducible research [@peng2011reproducible] can be tedious and error-prone using Hydroscope's web interface. On the contrary, using the Enhydris API for reproducibility requires external programs and scripting to import the data.
 
-`hydroscoper` provides functionality for automatic retrieval and translation of Hydroscope's data to English, for use in R [@Rbase]. The main functions that can be utilized is a family of functions, `get_stations`, `get_timeseries`, `get_data`, etc., to easily download JSON files to tidy data frames [@Wickham2014]. The internal databases of the package can be used to run queries on the available stations and time series, reducing the time needed for downloading and data wrangling [@kandel2011research], as these data are rarely modified.
+`hydroscoper` provides functionality for automatic retrieval and translation of Hydroscope's data to English, for use in R [@Rbase]. The main functions that can be utilized is the family of functions, `get_stations`, `get_timeseries`, `get_data`, etc., to easily download JSON and TXT files as tidy data frames [@Wickham2014]. The internal databases of the package can be used to run queries on the available stations and time series, reducing the time needed for downloading and data wrangling [@kandel2011research], as these data are rarely modified.
 
 The data have many applications. Availability of meteorological and hydrological information is essential for water resources management and environmental impact assessment studies [@vafiadis1994hydroscope]. They can be used by researchers to forward their studies when specific requirements of time series data are required, such as the estimation of rainfall erosivity [@vantas_sid2017]. Finally, these data are crucial to Greek organizations for the implementation of the Water Framework Directive 2000-2027 [@WFD].
 
