+++ data-raw/hydroscope_data.Rmd
@@ -199,7 +199,7 @@ timeseries <- ldply(subdomain, function(id){
   res <- merge(res, ts_units, by = "unit_of_measurement", all.x = TRUE)
   
   data.frame(
-    "timeser_id" = as.integer(res$id),
+    "time_id" = as.integer(res$id),
     "station_id" = as.integer(res$gentity),
     "variable" = as.character(res$variable_name),
     "timestep" = as.character(res$time_step_name),
