### 2018-02-21 20ee57c
#9 inconsistent usage of tibble/data frame/dataframe throughout documentation
use only the term `tibble` in documentation of all the datasets
