### 2018-02-21 a1bc00a
add   skip_on_travis() and  skip_on_appveyor() on test_enhydris.R
