### 2018-02-21 d7e9a3f
#4 add a vignette:  Introduction to Hydroscope(r)
