+++ vignettes/intro_hydroscoper.Rmd
@@ -1,0 +1,152 @@
+---
+title: "An introduction to `hydroscoper`"
+author: "Konstantinos Vantas"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## What is Hydroscope?
+
+[Hydroscope](http://www.hydroscope.gr/) is the  Greek National Data Bank for Hydrological and Meteorological Information, a result of long-standing efforts by numerous Greek scientists in collaboration with various companies and associations. It was implemented in three phases, funded by the Ministry of Development, the Ministry of Environment and Energy and the European Union.
+
+This National Data Bank provides several data sources from various organisations via a web interface. Each participating organisation keeps its data on its own server using a database system for the storage and management of hydrological and meteorological information. These organisations are:
+
+* Ministry of Environment and Energy.
+* Ministry of Rural Development and Food.
+* National Meteorological Service.
+* National Observatory of Athens.
+* Greek Prefectures.
+* Public Power Corporation.
+
+The above data are structured as tables and space separated text files, but are in Greek, thus limiting their usefulness. Another issue with Hydroscope is the lack of comprehensive look-up tables about the available data, which are spread across many different databases.  
+
+## What does `hydroscoper`?
+
+`hydroscoper` provides functionality for automatic retrieval and translation of Hydroscope's data to English.  The main functions that can be utilized is the family of functions, `get_stations`, `get_timeseries`, `get_data`, etc., to easily download Hydroscope's data as [tibbles](http://tibble.tidyverse.org/). 
+
+### Internal datasets
+
+The internal datasets of the package can be used to run queries on the available Hydroscope's stations and time series data, reducing the time needed for downloading and [data wrangling](http://bit.ly/1KslZb7), as these data are rarely modified.  These datasets are:
+
+#### `stations`
+
+It is a comprehensive look-up table with geographical and ownership information of the available stations in all Hydroscope's databases. The variables are: 
+
+  1. `station_id` The station's ID.
+  2. `name` The station's name.
+  3. `water_basin` The station's Water Basin.
+  4. `water_division` The station's Water Division.
+  5. `owner` The station's owner.
+  6. `longitude` The station's longitude in decimal degrees, [ETRS89](http://bit.ly/2kJwFuf).
+  7. `latitude` The station's latitude in decimal degrees, [ETRS89](http://bit.ly/2kJwFuf).
+  8. `altitude` The station's altitude, meters above sea level.
+  9. `subdomain` The corresponding Hydroscope's database.
+
+
+#### `timeseries`
+
+It is also a look-up table with all the available measurements for a given station in a  given Hydroscope's  database, with units of measurement and times of those measurements. The variables are:
+
+  1. `time_id` The time series ID.
+  2. `station_id` The corresponding station's ID.
+  3. `variable` The time series variable type.
+  4. `timestep` The timestep of time series.
+  5. `units` The units of the time series.
+  6. `start_date` The starting date of time series values.
+  7. `end_date` The ending date of time series values.
+  8. `subdomain` The corresponding Hydroscope's database.
+
+
+## Data sources
+
[…]
