### 2018-02-21 5636c7c
#10 further explanation about `timeseries`
1. Add explanation in README about both `timeseries` and `stations`
2. Add the above explanation in documentation
3.
