### 2018-02-21 709ae39
New version
* update version to 0.2.4
* add Sharla Gelfand as contributor
* update NEWS
* set Rproj Encoding to ASCII
* fix a typo in greece_borders documentation
