### 2018-02-21 #185 @kvantas
Hello again!  I have made the suggested changes:

# response to  @sharlagelfand

## Tests
 * The only file with UTF8 encoding is `README.md`, and it is created automatically. I 'll need some help on how to change it to ASCII.
 * On my laptop, for some strange reason, running `goodpractice::gp()` returns no errors  and I don't get a warning about LaTeX or encoding.
 * The skipped test is expected because subdomain `deh.hydroscope.gr` is currently down. Generally, if a server is down the corresponding test is skipped.

## Examples run unsuccessfully
I exported properly get_instruments_type() in https://github.com/kvantas/hydroscoper/commit/af0297e404b0a5e0b162afc96a0455194e98d9cc

## README
I added further explanation about `timeseries` in README, data set's documentation, and the vignettes in  https://github.com/kvantas/hydroscoper/commit/5636c7c84993960213eb28844afc0ae5957eb222

## Vignettes
 * I added a gentler introduction to Hydroscope and the hydroscoper package in https://github.com/kvantas/hydroscoper/commit/d7e9a3f1fa1376f70ca6520068a033d8b413fb6f
* Actually, `hydroscope_data` used to be a vignette, but I removed it due to its heavy web usage.

## Function Documentation

* In function: `get_data `: (1) I replaced `timeID` with `time_id`, (2) rewrote `@return` text, (3) replaced `stationID` with `time_id` at the comment and (4) used the term `tibble` instead of data-frame  in https://github.com/kvantas/hydroscoper/commit/445f0cc77b7405442a8b5b537857da2aeb48fd74
 * I used throughout documentation only the term tibble in https://github.com/kvantas/hydroscoper/commit/20ee57cd8aace1a869a3dc56fc76f11495bae817

## other recommendations

 * I renamed the variable name `timeser_id` in `timeseries` to `time_id` in https://github.com/kvantas/hydroscoper/commit/fdec5c5ee5b838b2749bff78f46085558c6e8c40 and in https://github.com/kvantas/hydroscoper/commit/e5b849e1d6973d20986aa58a710edf8c00dd8734
* I renamed `units` in `get_database()` list to `units_of_measurement` in https://github.com/kvantas/hydroscoper/commit/59f4acb34093ad9fff7d9470da8dfbd809c5c321
* I made `greece_borders` a tibble in https://github.com/kvantas/hydroscoper/commit/ca8840093e7fe370d30d9798f360e463ce21ef76
* I made `get_data()` function to return columns `date`, `value`, and `comment` in https://github.com/kvantas/hydroscoper/commit/a12b06a459bd3b8b7831c54027ec8574e826ea8f
