+++ .travis.yml
@@ -8,6 +8,10 @@ language: R
 sudo: false
 cache: packages
 
+env:
+  global:
+    - R_CHECK_ARGS="--no-build-vignettes --no-manual --as-cran rpack"
+
 r_packages:
   - covr
 
