### 2018-02-21 b9b0bad
use pingr::ping_port() function with port = 80 to test if a server is online
