+++ .travis.yml
@@ -8,9 +8,6 @@ language: R
 sudo: false
 cache: packages
 
-env:
-  not_cran: false
-  r_check_args: "--no-vignettes"
 
 r_packages:
   - covr
