### 2018-02-22 #185 @maelle
Thanks @kvantas!

@timtrice and @sharlagelfand could you please have a look and say whether you're satisfied with the changes made?
