### 2018-03-01 919e8d6
minor paper edits
* add hydroscoper in citation
* minor edits
