+++ paper.bib
@@ -81,8 +81,6 @@
   publisher={Sage Publications Sage UK: London, England}
 }
 
-
-
 @article{vantas_sid2017,
   author = {Vantas, Konstantinos and Sidiropoulos, Epameinontas},
   title = {Imputation of erosivity values under incomplete rainfall data by machine learning methods},
@@ -97,4 +95,15 @@
   author={Chow, VT and Maidment, DR and Mays, LW},
   year={1988},
   publisher={McGraw-Hill, New York}
+}
+
+@misc{hydroscoper2018,
+	title = {hydroscoper: R interface to the Greek National Data Bank for Hydrological and Meteorological Information},
+	shorttitle = {hydroscoper},
+	url = {https://github.com/kvantas/hydroscoper/},
+	urldate = {2018-03-01},
+	author = {Vantas, Konstantinos},
+	year = {2018},
+	keywords = {hydroscope, Greece, r, weather-data, water-resources},
+  Doi = {10.5281/zenodo.1182769}
 }
