### 2018-03-03 #185 @maelle
Approved! 👏

Thanks a lot @kvantas @sharlagelfand @timtrice for your work! :clap: :clap: :clap: Note that normally reviewers update the first checklist, but it's fine to have copied it, don't change it now! :smile_cat:

I have four *suggestions*:

* improving readability/browsability of the README via splitting it/adding a table of contents or even making it minimal and linking to the vignettes (instead of reproducing one of them).

* Add the pkgdown link to DESCRIPTION, see http://enpiar.com/2017/11/21/getting-down-with-pkgdown/ "if you already have a URL there (say, to your package’s GitHub repository), you can add a second URL to your pkgdown site, separated by a comma"

* If you wish you can add reviewers to DESCRIPTION with the "rev" role. If you do that ask them first though. :wink:

* In the pkgdown website you might want to make the grouping of functions in reference more useful. See https://github.com/dirkschumacher/ompr for an example

Now here is the list of things you have to do before I close this issue 😉

- [] Fix any links in badges for CI and coverage to point to the ropensci URL
- [] Submit to JOSS using the Zenodo DOI.  We will tag it for expedited review.

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.
