### 2018-03-03 #185 @kvantas
Thanks a lot @maelle, @timtrice and @sharlagelfan. @maelle, on the following days I will apply your suggestions and do what is needed to transfer the repo. Also, I will write a blog post about the package.

@timtrice do you mind if I add you as a reviewer?
@sharlagelfand is it ok to add you as a contributor/reviewer?
