### 2018-03-04 #185 @stefaniebutland
> I will write a blog post about the package
Excellent!
Here are some technical and editorial guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post

We ask that you submit your draft post via pull request a week before the planned publication date so we can give you some feedback.  I have a slot open on Tues April 3rd for a draft to be submitted March 27. How does that sound? If you submit earlier we can publish earlier if another is postponed
