### 2018-03-05 kvantas/hydroscoper#11 @kvantas
Improve readability/browsability of the README
Improving readability/browsability of the README via splitting it/adding a table of contents or even making it minimal and linking to the vignettes (instead of reproducing one of them).
