### 2018-03-05 kvantas/hydroscoper#12 @kvantas
Add the pkgdown link to DESCRIPTION
Add the pkgdown link to DESCRIPTION, see http://enpiar.com/2017/11/21/getting-down-with-pkgdown/ "if you already have a URL there (say, to your package’s GitHub repository), you can add a second URL to your pkgdown site, separated by a comma"
