### 2018-03-05 kvantas/hydroscoper#13 @kvantas
add reviewers to DESCRIPTION 
add reviewers to DESCRIPTION with the "rev" role
