### 2018-03-05 kvantas/hydroscoper#14 @kvantas
Make grouping of functions in reference more useful
In the pkgdown website you might want to make the grouping of functions in reference more useful. See https://github.com/dirkschumacher/ompr for an example
