### 2018-03-08 1919b0f
#13 add reviewers to DESCRIPTION
* add reviewers to DESCRIPTION with the "rev" role
* change title
