### 2018-03-08 f64ae3f
#14 make grouping of functions
* in pkgdown website use grouping of functions
