+++ _pkgdown.yml
@@ -1,0 +1,18 @@
+title: hydroscoper
+
+reference:
+  - title: Download tabular data
+    contents:
+    - get_database
+  - title: Download time series' values
+    contents:
+    - get_data
+  - title: Miscellaneous functions
+    contents:
+    - hydro_coords
+    - hydro_translate
+  - title: Datasets
+    contents:
+    - greece_borders
+    - stations
+    - timeseries
