### 2018-03-08 555efee
#12 Add the pkgdown link to DESCRIPTION
