+++ DESCRIPTION
@@ -17,7 +17,7 @@ Description: R interface to the  Greek National Data Bank for Hydrological and
     Meteorological Information <http://www.hydroscope.gr/>. It covers 
     Hydroscope's data sources and provides functions to transliterate, 
     translate and download them into tidy dataframes (tibbles).
-URL: https://github.com/kvantas/hydroscoper
+URL: https://github.com/kvantas/hydroscoper, https://kvantas.github.io/hydroscoper/
 BugReports: https://github.com/kvantas/hydroscoper/issues             
 Depends: R (>= 3.4.0)
 License: MIT + file LICENSE
