### 2018-03-10 5ef9c86
#11 make readme minimal
