+++ README.Rmd
@@ -26,70 +26,11 @@ knitr::opts_chunk$set(
 
 <img src="man/figures/hydroscoper_hex.png" align="right" height="220"/> 
 
-`hydroscoper` is an R interface to the  Greek National Data Bank for Hydrological and Meteorological Information, [Hydroscope](http://www.hydroscope.gr/). Hydroscope is the result of long-standing efforts by numerous Greek scientists in collaboration with various companies and associations. It was implemented in three phases, funded by the Ministry of Development, the Ministry of Environment and Energy and the European Union.
+`hydroscoper` is an R interface to the  Greek National Data Bank for Hydrological and Meteorological Information, [Hydroscope](http://www.hydroscope.gr/). For more details checkout the package's [ website](https://kvantas.github.io/hydroscoper/) and the
+vignettes:
 
-Hydroscope provides several national data sources from various organisations via a web interface. Each participating organisation keeps its data on its own server using the Enhydris database system for the storage and management of hydrological and meteorological data. These organisations are:
-
-* Ministry of Environment and Energy.
-* Ministry of Rural Development and Food.
-* National Meteorological Service.
-* National Observatory of Athens.
-* Greek Prefectures.
-* Public Power Corporation.
-
-The data are structured as tables and space separated text files, but are in Greek, thus limiting their usefulness. Another issue with Hydroscope is the lack of comprehensive look-up tables about the available data, which are spread across many different databases. 
-
-`hydroscoper` covers Hydroscope's data sources using the Enhydris API. The Enhydris database is implemented in PostgreSQL and details about the database can be found [here](http://bit.ly/2D0cZgA),  and about the Web-service API [here](http://bit.ly/2FlRtBB).
-
-`hydroscoper` provides functions to:
-
-1. Transform the available tables and data sets into tidy data frames [(tibbles)](http://tibble.tidyverse.org/).
-2. Transliterate the Greek Unicode text to Latin.
-3. Translate various Greek terms to English.
-
-The internal datasets of the package can be used to run queries on the available Hydroscope's stations and time series data, reducing the time needed for downloading and [data wrangling](http://bit.ly/1KslZb7), as these data are rarely modified.  These datasets are:
-
-### `stations`
-
-It is a comprehensive look-up table with geographical and ownership information of the available stations in all Hydroscope's databases. The variables are: 
-
-  1. `station_id` The station's ID.
-  2. `name` The station's name.
-  3. `water_basin` The station's Water Basin.
-  4. `water_division` The station's Water Division.
-  5. `owner` The station's owner.
-  6. `longitude` The station's longitude in decimal degrees, [ETRS89](http://bit.ly/2kJwFuf).
-  7. `latitude` The station's latitude in decimal degrees, [ETRS89](http://bit.ly/2kJwFuf).
-  8. `altitude` The station's altitude, meters above sea level.
-  9. `subdomain` The corresponding Hydroscope's database.
-
-
-### `timeseries`
-
-It is also a look-up table with all the available measurements for a given station in a  given Hydroscope's  database, with units of measurement and times of those measurements. The variables are:
-
-  1. `time_id` The time series ID.
-  2. `station_id` The corresponding station's ID.
-  3. `variable` The time series variable type.
-  4. `timestep` The timestep of time series.
-  5. `units` The units of the time series.
-  6. `start_date` The starting date of time series values.
[…]
