+++ vignettes/intro_hydroscoper.Rmd
@@ -20,7 +20,7 @@ knitr::opts_chunk$set(
 
 [Hydroscope](http://www.hydroscope.gr/) is the  Greek National Data Bank for Hydrological and Meteorological Information, a result of long-standing efforts by numerous Greek scientists in collaboration with various companies and associations. It was implemented in three phases, funded by the Ministry of Development, the Ministry of Environment and Energy and the European Union.
 
-This National Data Bank provides several data sources from various organisations via a web interface. Each participating organisation keeps its data on its own server using a database system for the storage and management of hydrological and meteorological information. These organisations are:
+This National Data Bank provides several data sources from various organisations via a web interface. Each participating organisation keeps its data on its own server using a database system for the storage and management of information. These organisations are:
 
 * Ministry of Environment and Energy.
 * Ministry of Rural Development and Food.
@@ -34,6 +34,8 @@ The above data are structured as tables and space separated text files, but are 
 ## What does `hydroscoper`?
 
 `hydroscoper` provides functionality for automatic retrieval and translation of Hydroscope's data to English.  The main functions that can be utilized is the family of functions, `get_stations`, `get_timeseries`, `get_data`, etc., to easily download Hydroscope's data as [tibbles](http://tibble.tidyverse.org/). 
+
+The package covers Hydroscope's data sources using the Enhydris API. The Enhydris database is implemented in PostgreSQL and details about the database can be found [here](http://bit.ly/2D0cZgA),  and about the Web-service API [here](http://bit.ly/2FlRtBB).
 
 ### Internal datasets
 
