### 2018-03-10 57fd506
#14 make grouping of functions in site more usuful
* create _pkgdown.yml
* rebuild site
