+++ paper.bib
@@ -105,5 +105,5 @@
 	author = {Vantas, Konstantinos},
 	year = {2018},
 	keywords = {hydroscope, Greece, r, weather-data, water-resources},
-  Doi = {10.5281/zenodo.1182769}
+  Doi = {http://doi.org/10.5281/zenodo.1182769}
 }
