+++ _pkgdown.yml
@@ -1,17 +1,21 @@
 title: hydroscoper
 
 reference:
-  - title: Download tabular data
+  - title: Download tables
+    desc: Funtions that retrieve tibbles with Hydroscope's data for a given data source
     contents:
     - get_database
-  - title: Download time series' values
+  - title: Download time series
+    desc: Retrieve a tibble with time series' values from Hydroscope
     contents:
     - get_data
   - title: Miscellaneous functions
+    desc: Internal functions and helpers
     contents:
     - hydro_coords
     - hydro_translate
   - title: Datasets
+    desc: The internal datasets of the package
     contents:
     - greece_borders
     - stations
