### 2018-03-12 ab240f6
minor edits
