+++ paper.bib
@@ -100,7 +100,7 @@
 @misc{hydroscoper2018,
 	title = {hydroscoper: R interface to the Greek National Data Bank for Hydrological and Meteorological Information},
 	shorttitle = {hydroscoper},
-	url = {https://github.com/kvantas/hydroscoper/},
+	url = {https://github.com/ropensci/hydroscoper/},
 	urldate = {2018-03-01},
 	author = {Vantas, Konstantinos},
 	year = {2018},
