+++ CONTRIBUTING.md
@@ -1,15 +1,15 @@
 # Suggestions and bug reports
 
-`hydroscoper` thrives on the suggestions and bug reports you submit to the [issue tracker](https://github.com/kvantas/hydroscoper/issues). Before posting, please search both the open and closed issues to help us avoid duplication.
+`hydroscoper` thrives on the suggestions and bug reports you submit to the [issue tracker](https://github.com/ropensci/hydroscoper/issues). Before posting, please search both the open and closed issues to help us avoid duplication.
 
 # Patches to code and documentation
 
-Please help us develop `hydroscoper`. It really helps us manage the exciting but accelerating stream of incoming [issues](https://github.com/kvantas/hydroscoper/issues). To submit a patch, please [fork this repository](https://help.github.com/articles/fork-a-repo/) and submit a [pull request](https://github.com/kvantas/hydroscoper/pulls).
+Please help us develop `hydroscoper`. It really helps us manage the exciting but accelerating stream of incoming [issues](https://github.com/ropensci/hydroscoper/issues). To submit a patch, please [fork this repository](https://help.github.com/articles/fork-a-repo/) and submit a [pull request](https://github.com/ropensci/hydroscoper/pulls).
 
 # Best ways to help
 
-Multiple [issues](https://github.com/kvantas/hydroscoper) are labeled ["help wanted"](https://github.com/kvantas/hydroscoper/issues?q=is%3Aissue+is%3Aopen+label%3A%22help+wanted%22). We would particularly appreciate your input there.
+Multiple [issues](https://github.com/ropensci/hydroscoper) are labeled ["help wanted"](https://github.com/ropensci/hydroscoper/issues?q=is%3Aissue+is%3Aopen+label%3A%22help+wanted%22). We would particularly appreciate your input there.
 
 # Contributing
 
-Maintainers and other contributors must follow this repository's [code of conduct](https://github.com/kvantas/hydroscoper/blob/master/CONDUCT.md).
+Maintainers and other contributors must follow this repository's [code of conduct](https://github.com/ropensci/hydroscoper/blob/master/CONDUCT.md).
