### 2018-03-12 #185 @kvantas
@maelle, sorry for the delay, some work kept me back. I followed all your suggestions in
https://github.com/ropensci/hydroscoper/issues/11,
https://github.com/ropensci/hydroscoper/issues/12,
https://github.com/ropensci/hydroscoper/issues/13,
https://github.com/ropensci/hydroscoper/issues/14

But I have some problems:

a) ropensci/hydroscoper in travis is locked
b) I cannot edit the description in https://github.com/ropensci/hydroscoper
c) I cannot find a link in Zenodo
