### 2018-03-12 b57db45
fix github install chunk in README
