+++ README.Rmd
@@ -44,7 +44,7 @@ You can install the development version from GitHub with:
 
 ```{r gh-installation, eval = FALSE}
 # install.packages("devtools")
-devtools::install_github("kvantas/hydroscoper")
+devtools::install_github("ropensci/hydroscoper")
 ```
 
 ## Using hydroscoper
