### 2018-03-12 90b0117
remove redundant text and chunk in README
