### 2018-03-12 #185 @maelle
a) yeah you do not have access, what do you need to do? I'll be traveling soon so please ping one of the editors in the # onboarding Slack channel
b) c) I made you an admin of the repo now, which I couldn't do before transfer 😉
