### 2018-03-12 #185 @maelle
d) Appveyor now activated
