### 2018-03-14 #185 @kvantas
@maelle, I submitted the :package: to JOSS: https://github.com/openjournals/joss-reviews/issues/621 , the last thing on the list!
