### 2018-03-20 #185 @stefaniebutland
@kvantas Are you still ok with submitting your draft blog post by Tues Mar 27?
